Modified the type to bigint for code_num field in codes table
The type was giving convertion problem when providing code with some number in it that could be above a bigint.
It caused problem at some lines to be imported.
Thanks to Franck Theeten for the solution ;)

fix #1794

+++ data/db/createfunctions.sql
@@ -122,7 +122,7 @@ BEGIN
                 IF codeNum = '' THEN
                   NEW.code_num := 0;
                 ELSE
-                  NEW.code_num := codeNum::int;
+                  NEW.code_num := codeNum::bigint;
                 END IF;
                 NEW.full_code_indexed := fullToIndex(COALESCE(NEW.code_prefix,'') || COALESCE(NEW.code::text,'') || COALESCE(NEW.code_suffix,'') );
         ELSIF TG_TABLE_NAME = 'tag_groups' THEN
